+++ app/builders/messages/instagram/message_builder.rb
@@ -1,4 +1,8 @@
 class Messages::Instagram::MessageBuilder < Messages::Instagram::BaseMessageBuilder
+  include HTTParty
+
+  base_uri "https://graph.instagram.com/#{GlobalConfigService.load('INSTAGRAM_API_VERSION', 'v22.0')}"
+
   def initialize(messaging, inbox, outgoing_echo: false)
     super(messaging, inbox, outgoing_echo: outgoing_echo)
   end
@@ -6,28 +10,33 @@ class Messages::Instagram::MessageBuilder < Messages::Instagram::BaseMessageBuil
   private
 
   def get_story_object_from_source_id(source_id)
-    k = Koala::Facebook::API.new(@inbox.channel.page_access_token) if @inbox.facebook?
-    k.get_object(source_id, fields: %w[story from]) || {}
-  rescue Koala::Facebook::AuthenticationError
-    @inbox.channel.authorization_error!
-    raise
-  rescue Koala::Facebook::ClientError => e
-    # The exception occurs when we are trying fetch the deleted story or blocked story.
-    @message.attachments.destroy_all
-    @message.update(content: I18n.t('conversations.messages.instagram_deleted_story_content'))
-    Rails.logger.error e
+    url = "#{self.class.base_uri}/#{source_id}?fields=story,from&access_token=#{@inbox.channel.access_token}"
+    response = HTTParty.get(url)
+
+    Rails.logger.info("Instagram Story Response: #{response.body}")
+
+    return JSON.parse(response.body).with_indifferent_access if response.success?
+
+    handle_error_response(response)
     {}
   rescue StandardError => e
-    ChatwootExceptionTracker.new(e, account: @inbox.account).capture_exception
+    handle_standard_error(e)
     {}
   end
 
-  def find_conversation_scope
-    Conversation.where(conversation_params)
-                .where("additional_attributes ->> 'type' = 'instagram_direct_message'")
+  def handle_error_response(response)
+    return unless response.code == 404
+
+    @message.attachments.destroy_all
+    @message.update(content: I18n.t('conversations.messages.instagram_deleted_story_content'))
   end
 
-  def additional_conversation_attributes
-    { type: 'instagram_direct_message' }
+  def handle_standard_error(error)
+    if error.response&.unauthorized?
+      @inbox.channel.authorization_error!
+      raise
+    end
+    Rails.logger.error("Instagram Story Error: #{error.message}")
+    ChatwootExceptionTracker.new(error, account: @inbox.account).capture_exception
   end
 end
