+++ app/jobs/send_reply_job.rb
@@ -13,7 +13,7 @@ class SendReplyJob < ApplicationJob
       'Channel::Telegram' => ::Telegram::SendOnTelegramService,
       'Channel::Whatsapp' => ::Whatsapp::SendOnWhatsappService,
       'Channel::Sms' => ::Sms::SendOnSmsService,
-      'Channel::Instagram' => ::Instagram::Direct::SendOnInstagramService
+      'Channel::Instagram' => ::Instagram::SendOnInstagramService
     }
 
     case channel_name
@@ -28,7 +28,7 @@ class SendReplyJob < ApplicationJob
 
   def send_on_facebook_page(message)
     if message.conversation.additional_attributes['type'] == 'instagram_direct_message'
-      ::Instagram::SendOnInstagramService.new(message: message).perform
+      ::Instagram::Messenger::SendOnInstagramService.new(message: message).perform
     else
       ::Facebook::SendOnFacebookService.new(message: message).perform
     end
