--- app/services/instagram/direct/send_on_instagram_service.rb
@@ -1,33 +1,0 @@
-class Instagram::Direct::SendOnInstagramService < Instagram::BaseSendService
-  private
-
-  def channel_class
-    Channel::Instagram
-  end
-
-  # Deliver a message with the given payload.
-  # https://developers.facebook.com/docs/instagram-platform/instagram-api-with-instagram-login/messaging-api
-  def send_message(message_content)
-    access_token = channel.access_token
-    query = { access_token: access_token }
-    instagram_id = channel.instagram_id.presence || 'me'
-
-    response = HTTParty.post(
-      "https://graph.instagram.com/v22.0/#{instagram_id}/messages",
-      body: message_content,
-      query: query
-    )
-
-    process_response(response, message_content)
-  end
-
-  def merge_human_agent_tag(params)
-    global_config = GlobalConfig.get('ENABLE_INSTAGRAM_CHANNEL_HUMAN_AGENT')
-
-    return params unless global_config['ENABLE_INSTAGRAM_CHANNEL_HUMAN_AGENT']
-
-    params[:messaging_type] = 'MESSAGE_TAG'
-    params[:tag] = 'HUMAN_AGENT'
-    params
-  end
-end
