+++ app/services/instagram/send_on_instagram_service.rb
@@ -2,19 +2,18 @@ class Instagram::SendOnInstagramService < Instagram::BaseSendService
   private
 
   def channel_class
-    Channel::FacebookPage
+    Channel::Instagram
   end
 
   # Deliver a message with the given payload.
-  # @see https://developers.facebook.com/docs/messenger-platform/instagram/features/send-message
+  # https://developers.facebook.com/docs/instagram-platform/instagram-api-with-instagram-login/messaging-api
   def send_message(message_content)
-    access_token = channel.page_access_token
-    app_secret_proof = calculate_app_secret_proof(GlobalConfigService.load('FB_APP_SECRET', ''), access_token)
+    access_token = channel.access_token
     query = { access_token: access_token }
-    query[:appsecret_proof] = app_secret_proof if app_secret_proof
+    instagram_id = channel.instagram_id.presence || 'me'
 
     response = HTTParty.post(
-      'https://graph.facebook.com/v11.0/me/messages',
+      "https://graph.instagram.com/v22.0/#{instagram_id}/messages",
       body: message_content,
       query: query
     )
@@ -22,16 +21,10 @@ class Instagram::SendOnInstagramService < Instagram::BaseSendService
     process_response(response, message_content)
   end
 
-  def calculate_app_secret_proof(app_secret, access_token)
-    Facebook::Messenger::Configuration::AppSecretProofCalculator.call(
-      app_secret, access_token
-    )
-  end
-
   def merge_human_agent_tag(params)
-    global_config = GlobalConfig.get('ENABLE_MESSENGER_CHANNEL_HUMAN_AGENT')
+    global_config = GlobalConfig.get('ENABLE_INSTAGRAM_CHANNEL_HUMAN_AGENT')
 
-    return params unless global_config['ENABLE_MESSENGER_CHANNEL_HUMAN_AGENT']
+    return params unless global_config['ENABLE_INSTAGRAM_CHANNEL_HUMAN_AGENT']
 
     params[:messaging_type] = 'MESSAGE_TAG'
     params[:tag] = 'HUMAN_AGENT'
