+++ spec/services/instagram/send_on_instagram_service_spec.rb
@@ -3,24 +3,20 @@ require 'rails_helper'
 describe Instagram::SendOnInstagramService do
   subject(:send_reply_service) { described_class.new(message: message) }
 
-  before do
-    stub_request(:post, /graph\.facebook\.com/)
-    create(:message, message_type: :incoming, inbox: instagram_inbox, account: account, conversation: conversation)
-  end
-
   let!(:account) { create(:account) }
-  let!(:instagram_channel) { create(:channel_instagram_fb_page, account: account, instagram_id: 'chatwoot-app-user-id-1') }
-  let!(:instagram_inbox) { create(:inbox, channel: instagram_channel, account: account, greeting_enabled: false) }
+  let!(:instagram_direct_channel) { create(:channel_instagram, account: account, instagram_id: 'instagram-message-id-123') }
+  let!(:instagram_direct_inbox) { create(:inbox, channel: instagram_direct_channel, account: account, greeting_enabled: false) }
+
   let!(:contact) { create(:contact, account: account) }
-  let(:contact_inbox) { create(:contact_inbox, contact: contact, inbox: instagram_inbox) }
-  let(:conversation) { create(:conversation, contact: contact, inbox: instagram_inbox, contact_inbox: contact_inbox) }
+  let(:contact_inbox) { create(:contact_inbox, contact: contact, inbox: instagram_direct_inbox) }
+  let(:conversation) { create(:conversation, contact: contact, inbox: instagram_direct_inbox, contact_inbox: contact_inbox) }
   let(:response) { double }
   let(:mock_response) do
     instance_double(
       HTTParty::Response,
       :success? => true,
-      :body => { message_id: 'anyrandommessageid1234567890' }.to_json,
-      :parsed_response => { 'message_id' => 'anyrandommessageid1234567890' }
+      :body => { message_id: 'random_message_id' }.to_json,
+      :parsed_response => { 'message_id' => 'random_message_id' }
     )
   end
 
@@ -56,24 +52,24 @@ describe Instagram::SendOnInstagramService do
   describe '#perform' do
     context 'with reply' do
       before do
-        allow(Facebook::Messenger::Configuration::AppSecretProofCalculator).to receive(:call).and_return('app_secret_key', 'access_token')
         allow(HTTParty).to receive(:post).and_return(mock_response)
       end
 
       context 'without message_tag HUMAN_AGENT' do
         before do
-          InstallationConfig.where(name: 'ENABLE_MESSENGER_CHANNEL_HUMAN_AGENT').first_or_create(value: false)
+          InstallationConfig.where(name: 'ENABLE_INSTAGRAM_CHANNEL_HUMAN_AGENT').first_or_create(value: false)
         end
 
         it 'if message is sent from chatwoot and is outgoing' do
-          message = create(:message, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
+          message = create(:message, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account, conversation: conversation)
 
           response = described_class.new(message: message).perform
-          expect(response['message_id']).to eq('anyrandommessageid1234567890')
+          expect(response['message_id']).to eq('random_message_id')
         end
 
         it 'if message is sent from chatwoot and is outgoing with multiple attachments' do
-          message = build(:message, content: nil, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
+          message = build(:message, content: nil, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account,
+                                    conversation: conversation)
           avatar = message.attachments.new(account_id: message.account_id, file_type: :image)
           avatar.file.attach(io: Rails.root.join('spec/assets/avatar.png').open, filename: 'avatar.png', content_type: 'image/png')
           sample = message.attachments.new(account_id: message.account_id, file_type: :image)
@@ -91,17 +87,17 @@ describe Instagram::SendOnInstagramService do
         end
 
         it 'if message with attachment is sent from chatwoot and is outgoing' do
-          message = build(:message, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
+          message = build(:message, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account, conversation: conversation)
           attachment = message.attachments.new(account_id: message.account_id, file_type: :image)
           attachment.file.attach(io: Rails.root.join('spec/assets/avatar.png').open, filename: 'avatar.png', content_type: 'image/png')
           message.save!
           response = described_class.new(message: message).perform
 
-          expect(response['message_id']).to eq('anyrandommessageid1234567890')
+          expect(response['message_id']).to eq('random_message_id')
         end
 
         it 'if message sent from chatwoot is failed' do
-          message = create(:message, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
+          message = create(:message, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account, conversation: conversation)
 
           allow(HTTParty).to receive(:post).and_return(response_with_error)
           described_class.new(message: message).perform
@@ -117,11 +113,11 @@ describe Instagram::SendOnInstagramService do
         end
 
         it 'if message is sent from chatwoot and is outgoing' do
-          message = create(:message, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
+          message = create(:message, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account, conversation: conversation)
 
           allow(HTTParty).to receive(:post).with(
             {
-              recipient: { id: contact.get_source_id(instagram_inbox.id) },
+              recipient: { id: contact.get_source_id(instagram_direct_inbox.id) },
               message: {
                 text: message.content
               },
@@ -130,7 +126,7 @@ describe Instagram::SendOnInstagramService do
             }
           ).and_return(
             {
-              'message_id': 'anyrandommessageid1234567890'
+              'message_id': 'random_message_id'
             }
           )
 
@@ -138,39 +134,35 @@ describe Instagram::SendOnInstagramService do
           expect(HTTParty).to have_received(:post)
         end
       end
-    end
 
-    context 'when handling errors' do
-      before do
-        allow(Facebook::Messenger::Configuration::AppSecretProofCalculator).to receive(:call).and_return('app_secret_key', 'access_token')
-      end
+      context 'when handling errors' do
+        it 'handles HTTP errors' do
+          message = create(:message, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account, conversation: conversation)
+          allow(HTTParty).to receive(:post).and_return(error_response)
 
-      it 'handles HTTP errors' do
-        message = create(:message, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
-        allow(HTTParty).to receive(:post).and_return(error_response)
+          described_class.new(message: message).perform
 
-        described_class.new(message: message).perform
+          expect(message.reload.status).to eq('failed')
+          expect(message.reload.external_error).to eq('400 - The Instagram account is restricted.')
+        end
 
-        expect(message.reload.status).to eq('failed')
-        expect(message.reload.external_error).to eq('400 - The Instagram account is restricted.')
-      end
+        it 'handles response errors' do
+          message = create(:message, message_type: 'outgoing', inbox: instagram_direct_inbox, account: account, conversation: conversation)
 
-      it 'handles response errors' do
-        message = create(:message, message_type: 'outgoing', inbox: instagram_inbox, account: account, conversation: conversation)
+          error_response = instance_double(
+            HTTParty::Response,
+            success?: true,
+            body: { 'error' => { 'message' => 'Invalid message format', 'code' => 100 } }.to_json,
+            parsed_response: { 'error' => { 'message' => 'Invalid message format', 'code' => 100 } }
+          )
 
-        error_response = instance_double(
-          HTTParty::Response,
-          success?: true,
-          body: { 'error' => { 'message' => 'Invalid message format', 'code' => 100 } }.to_json,
-          parsed_response: { 'error' => { 'message' => 'Invalid message format', 'code' => 100 } }
-        )
+          allow(HTTParty).to receive(:post).and_return(error_response)
 
-        allow(HTTParty).to receive(:post).and_return(error_response)
+          described_class.new(message: message).perform
 
-        described_class.new(message: message).perform
-
-        expect(message.reload.status).to eq('failed')
-        expect(message.reload.external_error).to eq('100 - Invalid message format')
+          expect(message.reload.status).to eq('failed')
+          expect(message.reload.external_error).to eq('100 - Invalid message format')
+        end
       end
     end
   end
